fix: cleaned up.

+++ kernel/include/arch/bus/pci/device.h
@@ -30,7 +30,7 @@
 
 struct PCIDevice
 {
-    uint16_t bus;
+    uint8_t bus;
     uint8_t slot;
     uint8_t func;
     uint8_t valid : 1;
